+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Removed support for Go 1.9.
     * Added `Rows` and `GroupBy` calls.
     * Added roaring import support for `RowKeyColumnID`, `RowIDColumnKey` and `RowKeyColumnKey` type data. Pass `OptImportRoaring(true)` to `client.ImportField` to activate that.
+    * Deprecated `Range` call. Use `RowRange` instead.
 
 * **v1.2.0** (2018-12-21)
     * **Compatible with Pilosa 1.2**